feat(workflow): allow creating inbox from workflow
