+++ packages/app-builder/public/locales/en/workflows.json
@@ -32,12 +32,11 @@
   "detail_panel.decision_created.outcomes.label": "Outcomes",
   "detail_panel.decision_created.outcomes.placeholder": "Select at least one outcome",
   "detail_panel.create_case.description": "Create a new case with the created decision. The case will be created in the selected inbox with a default name.",
-  "detail_panel.create_case.inbox.label": "Inbox",
-  "detail_panel.create_case.inbox.placeholder": "Select an inbox",
+  "detail_panel.inbox.label": "Inbox",
+  "detail_panel.inbox.placeholder": "Select an inbox",
+  "detail_panel.inbox.need_inbox_contact_admin": "Please contact your admin to create an inbox",
   "detail_panel.create_case.default_name.helper": "The default name should follow this template:",
   "detail_panel.add_to_case_if_possible.description": "In the selected inbox, if an open case matches the decision’s pivot value, add the decision to it. If no matching case is found, a new case will be created with a default name.",
-  "detail_panel.add_to_case_if_possible.inbox.label": "Inbox",
-  "detail_panel.add_to_case_if_possible.inbox.placeholder": "Select an inbox",
   "detail_panel.add_to_case_if_possible.default_name.helper": "If no matching case is found, the created case name should follow this template:",
   "detail_panel.checklist.title": "Checklist",
   "detail_panel.checklist.description": "Make sure all issues are resolved before saving the workflow.",
