+++ packages/app-builder/src/components/Scenario/Workflow/DetailPanel/DetailPanel.tsx
@@ -47,7 +47,8 @@ export function DetailPanel({ onSave, onDelete }: DetailPanelProps) {
   };
 
   return (
-    <div className="border-grey-10 bg-grey-00 flex h-full flex-col overflow-hidden border-l">
+    // eslint-disable-next-line tailwindcss/no-custom-classname -- nokey class ensures that ReactFlow onKeyDown event is not triggered (ex: delete node)
+    <div className="border-grey-10 bg-grey-00 nokey flex h-full flex-col overflow-hidden border-l">
       <ScrollAreaV2 type="auto" className="flex-1">
         <div className="flex h-full flex-col gap-4 p-6">
           <DetailPanelContent />
