+++ packages/app-builder/src/components/Scenario/Workflow/workflow-i18n.ts
@@ -4,4 +4,5 @@ export const workflowI18n = [
   'common',
   'workflows',
   'decisions',
+  'settings', // Required by <CreateInbox />
 ] satisfies Namespace;
