+++ packages/app-builder/src/routes/ressources+/settings+/inboxes+/create.tsx
@@ -13,7 +13,7 @@ import { useFetcher, useNavigation } from '@remix-run/react';
 import { type Namespace } from 'i18next';
 import { useEffect, useId, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Button, Modal } from 'ui-design-system';
+import { Button, ModalV2 } from 'ui-design-system';
 import { Icon } from 'ui-icons';
 import { z } from 'zod';
 
@@ -27,7 +27,7 @@ export const redirectRouteOptions = [
 ] as const;
 const createInboxFormSchema = z.object({
   name: z.string().min(1),
-  redirectRoute: z.enum(redirectRouteOptions),
+  redirectRoute: z.enum(redirectRouteOptions).optional(),
 });
 
 export async function action({ request }: ActionFunctionArgs) {
@@ -43,18 +43,20 @@ export async function action({ request }: ActionFunctionArgs) {
   const submission = parse(formData, { schema: createInboxFormSchema });
 
   if (submission.intent !== 'submit' || !submission.value) {
-    return json(submission);
+    return json({ submission, success: false });
   }
 
   try {
     const { inbox: createdInbox } = await apiClient.createInbox({
       name: submission.value.name,
     });
-    return redirect(
-      getRoute(submission.value.redirectRoute, {
-        inboxId: fromUUID(createdInbox.id),
-      }),
-    );
+    if (submission.value.redirectRoute)
+      return redirect(
+        getRoute(submission.value.redirectRoute, {
+          inboxId: fromUUID(createdInbox.id),
+        }),
+      );
+    return json({ submission, success: true });
   } catch (error) {
     const session = await getSession(request);
 
@@ -63,16 +65,19 @@ export async function action({ request }: ActionFunctionArgs) {
       messageKey: 'common:errors.unknown',
     });
 
-    return json(submission, {
-      headers: { 'Set-Cookie': await commitSession(session) },
-    });
+    return json(
+      { submission, success: false },
+      {
+        headers: { 'Set-Cookie': await commitSession(session) },
+      },
+    );
   }
 }
 
 export function CreateInbox({
   redirectRoutePath,
 }: {
-  redirectRoutePath: (typeof redirectRouteOptions)[number];
+  redirectRoutePath?: (typeof redirectRouteOptions)[number];
 }) {
   const { t } = useTranslation(handle.i18n);
   const [open, setOpen] = useState(false);
@@ -85,38 +90,45 @@ export function CreateInbox({
   }, [navigation.state]);
 
   return (
-    <Modal.Root open={open} onOpenChange={setOpen}>
-      <Modal.Trigger asChild>
-        <Button
-          className="whitespace-nowrap"
-          variant="secondary"
-          onClick={(e) => e.stopPropagation()}
-        >
-          <Icon icon="new-inbox" className="size-5 shrink-0" />
-          {t('settings:inboxes.new_inbox.create')}
-        </Button>
-      </Modal.Trigger>
-      <Modal.Content onClick={(e) => e.stopPropagation()}>
-        <CreateInboxContent redirectRoutePath={redirectRoutePath} />
-      </Modal.Content>
-    </Modal.Root>
+    <ModalV2.Root open={open} setOpen={setOpen}>
+      <ModalV2.Trigger
+        onClick={(e) => e.stopPropagation()}
+        render={<Button className="whitespace-nowrap" variant="secondary" />}
+      >
+        <Icon icon="new-inbox" className="size-5 shrink-0" />
+        {t('settings:inboxes.new_inbox.create')}
+      </ModalV2.Trigger>
+      <ModalV2.Content onClick={(e) => e.stopPropagation()}>
+        <CreateInboxContent
+          redirectRoutePath={redirectRoutePath}
+          setOpen={setOpen}
+        />
+      </ModalV2.Content>
+    </ModalV2.Root>
   );
 }
 
 export function CreateInboxContent({
   redirectRoutePath,
+  setOpen,
 }: {
-  redirectRoutePath: (typeof redirectRouteOptions)[number];
+  redirectRoutePath?: (typeof redirectRouteOptions)[number];
+  setOpen: (open: boolean) => void;
 }) {
   const { t } = useTranslation(handle.i18n);
 
   const fetcher = useFetcher<typeof action>();
+  useEffect(() => {
+    if (fetcher?.data?.success) {
+      setOpen(false);
+    }
+  }, [setOpen, fetcher?.data?.success]);
 
   const formId = useId();
   const [form, { name, redirectRoute }] = useForm({
     id: formId,
     defaultValue: { name: '', redirectRoute: redirectRoutePath },
-    lastSubmission: fetcher.data,
+    lastSubmission: fetcher.data?.submission,
     constraint: getFieldsetConstraint(createInboxFormSchema),
     onValidate({ formData }) {
       return parse(formData, {
@@ -131,7 +143,7 @@ export function CreateInboxContent({
       action={getRoute('/ressources/settings/inboxes/create')}
       {...form.props}
     >
-      <Modal.Title>{t('settings:inboxes.new_inbox.explain')}</Modal.Title>
+      <ModalV2.Title>{t('settings:inboxes.new_inbox.explain')}</ModalV2.Title>
       <div className="flex flex-col gap-6 p-6">
         <input {...conform.input(redirectRoute, { type: 'hidden' })} />
         <FormField config={name} className="group flex flex-col gap-2">
@@ -140,11 +152,11 @@ export function CreateInboxContent({
           <FormError />
         </FormField>
         <div className="flex flex-1 flex-row gap-2">
-          <Modal.Close asChild>
-            <Button className="flex-1" variant="secondary">
-              {t('common:cancel')}
-            </Button>
-          </Modal.Close>
+          <ModalV2.Close
+            render={<Button className="flex-1" variant="secondary" />}
+          >
+            {t('common:cancel')}
+          </ModalV2.Close>
           <Button
             className="flex-1"
             variant="primary"
